+++ packages/web/src/controls/Control.tsx
@@ -107,6 +107,7 @@ const ControlWithRef = forwardRef(function ControlWithRef<T extends string>(
     borderRadius,
     borderWidth,
     color = checked || indeterminate ? 'fg' : 'fgMuted',
+    elevation,
     testID,
     iconStyle,
     labelStyle,
@@ -138,6 +139,7 @@ const ControlWithRef = forwardRef(function ControlWithRef<T extends string>(
         borderWidth={borderWidth}
         className={interactableCss}
         disabled={disabled || readOnly}
+        elevation={elevation}
         style={iconStyle}
         testID={testID ? `${testID}-parent` : undefined}
       >
@@ -169,6 +171,7 @@ const ControlWithRef = forwardRef(function ControlWithRef<T extends string>(
       checked,
       children,
       disabled,
+      elevation,
       htmlProps,
       iconStyle,
       inputId,